+++ src/features/giveaways/web/send.rs
@@ -25,7 +25,7 @@ pub async fn handle_send_giveaway_message(
     Path((guild_id_str, config_id_str)): Path<(String, String)>,
 ) -> Result<(StatusCode, Json<SendGiveawayResponse>), (StatusCode, String)> {
     let config_id = parse_config_id(&config_id_str)?;
-    let guild_id: i64 = guild_id_str.parse().map_err(|e| {
+    let guild_id: u64 = guild_id_str.parse().map_err(|e| {
         warn!(error = ?e, guild_id_str, "Invalid guild_id format");
         (StatusCode::BAD_REQUEST, "Invalid guild ID".to_string())
     })?;
@@ -46,7 +46,7 @@ pub async fn handle_send_giveaway_message(
         .inspect_err(|e| warn!(error = ?e, "Couldn't get user through HTTP"))
         .map_err(|_| (StatusCode::INTERNAL_SERVER_ERROR, "Internal server error.".to_string()))?;
 
-    let gctx = get_guild_ctx(GuildId::from(guild_id as u64), &state.serenity_http)
+    let gctx = get_guild_ctx(GuildId::from(guild_id), &state.serenity_http)
         .await
         .inspect_err(|e| warn!(error = ?e, "Couldn't get guild ctx"))
         .map_err(|_| (StatusCode::INTERNAL_SERVER_ERROR, "Internal server error.".to_string()))?;
